requirements.txt
fastapi
uvicorn
torch
transformers
pydantic-settings
python-dotenv
pandas
scikit-learn
wandb
onnx
onnxruntime-gpu
numpy